Add benchmark functionality and correctness checks
This commit adds support for timing the iterations of a model being
run by torchbench.py. For multiple iterations, the script also outputs
the mean time and standard deviation. This commit also adds a simple
correctness check of the output from the torch-mlir compiled module
against the eager equivalent in pytorch.