Add CUDA LLVM target arch flag to `compile_to_vmfb`
In order to get the best performance on CUDA, the CUDA LLVM target
architecture must be specified in the IREE compiler. This commit adds
support for specifying such a flag. In addition, it sets the flag to
`sm_80` in the TorchDynamo `utils.py`, since that target architecture
is needed to use NVIDIA's tensorcore.